// src/types/DetailType.ts
export type DetailType = {
  organizationNumber: string;
  organizationName: string;
  organizationFormDescription: string;
  registrationDate: string;
  numberOfEmployee: number;
  country: string;
  countryCode: string;
  postCode: string;
  postOffice: string;
  address: Array<string>;
  municipality: string;
  municipalityNumber: string;
  establishedDate: string;
  institutionSectorCode: string;
  institutionSectorDescription: string;
  bankruptcy: boolean;
  maalform: string;
  sourceUrl: string;
};